take a uint8_t instead of a size_t in random_bytes
Since this function assumes the caller will only ever pass in ≤256, we may as
well encode this in the type system. This also avoids an unpleasant case where a
user could cause an assertion failure now that this is part of the public API.

+++ include/passwand/passwand.h
@@ -188,4 +188,4 @@ void passwand_secure_heap_print(FILE *f);
  * @param buffer_len  Number of bytes requested
  * @return            PW_OK on success
  */
-passwand_error_t passwand_random_bytes(void *buffer, size_t buffer_len);
+passwand_error_t passwand_random_bytes(void *buffer, uint8_t buffer_len);
